fix: model extraction
Due to inlining in Spacer, sometimes we cannot extract Counterex from the output. One may resort to using query!0. However, we cannot guarantee that the order of arguments in query!0 are the same as those in Counterex. Therefore, we have to disable inlining.

Not that some cases involves nonlinear operators like `*`. We need to use a different arithmetic solver when handling these.